:recycle:  move graphql modules under graphql folder
