+++ packages/utils/src/index.ts
@@ -4,7 +4,8 @@ export * from "./array";
 export * from "./object";
 export * from "./string";
 export * as pathUrl from "./url";
-export * from "./graphql";
-export * from "./group";
-export * from "./directive";
+export * from "./graphql/graphql";
+export * from "./graphql/group";
+export * from "./graphql/directive";
 export * from "./logger";
+export * from "./graphql/formatter";
